+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 Version 0.5.2
 -------------
 - Cleaned up build solvent mask calculation
+- Added DeviceMixin for centralizing device handling accross all classes
 
 Version 0.5.1
 -------------